+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
+- (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService
 
 ## [1.4.0](https://github.com/arangodb/kube-arangodb/tree/1.4.0) (2026-01-20)
 - (Feature) Deprecate Dashboard